Phase 4: decompose CreatorStudioView into eight modules
CreatorStudioView.tsx was 2,247 lines: the orchestrator plus every panel
it renders, seven sub-components deep. Split the whole tail of the file
out in one pass, leaving the orchestrator alone in it.

New modules, all previously in-file with no external consumers:

- studioTabs.ts        tab resolution + URL addressing
- studioHealth.ts      formatSeconds + health-row lookup
- StudioShelf.tsx      shelf search/filter controls + the game list
- StudioDetailsPanel.tsx  the icon-rail details panel and its pane type
- DraftShareControl.tsx   the draft-share switch and permalink
- StudioStatsSection.tsx  health numbers + player reactions
- StudioSuggestions.tsx   the per-game suggestion inbox
- AutonomySetting.tsx     the autonomy permission scale

CreatorStudioView.tsx drops to 1,192 lines and its imports shrink to what
the orchestrator itself uses. Prose comments on the moved code are
rewritten as one-line notes so each new file starts at zero debt; the
module-size and comment-prose baselines for CreatorStudioView.tsx are
ratcheted down to the new figures rather than left at the old ceiling.

No behaviour change: same components, same props, same render order. One
narrowing detail moved — the contributions block now reads
`showContributions && game.slug` so the slug stays narrowed for
ContributionsSetting.
