+++ apps/web/src/surfaces/studio/CreatorStudioView.tsx
@@ -1,15 +1,13 @@
-import type { DismissReason } from '@gamedevpl/contract';
-import { useEffect, useId, useMemo, useRef, useState, type RefObject } from 'react';
+import { useEffect, useId, useMemo, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useAuth } from '../../AuthContext.js';
 import { AuthModal } from '../../AuthModal.js';
 import { ClaimHandleModal } from '../../ClaimHandleModal.js';
 import { StudioCreatorProfileProvider } from '../../studioCreatorProfile.js';
 import type { GameHealth } from '../../healthApi.js';
-import { PixelIcon, type PixelIconName } from '../../PixelIcon.js';
-import { formatRelativeTime } from '../../relativeTime.js';
-import { playPath, studioPath, type StudioTab } from '../../core/router.js';
-import { abandonSubmission, deleteGame, handoffToPlatform } from '../../submissionApi.js';
+import { PixelIcon } from '../../PixelIcon.js';
+import { studioPath, type StudioTab } from '../../core/router.js';
+import { handoffToPlatform } from '../../submissionApi.js';
 import { StudioShotToasts } from './StudioShotToasts.js';
 import { type CodeActionsMode } from './CodeActionsMenu.js';
 import { CodeSurface } from './CodeSurface.js';
@@ -30,35 +28,23 @@ import {
   isStudioGamePublished,
   isStudioGameShelfLive,
   sortStudioGames,
-  studioGameInitials,
   STUDIO_LIVE_STATUSES,
   STUDIO_SHELF_TOOLS_AT,
   type StudioShelfFilter,
   type StudioShelfGame,
 } from '../../studioShelf.js';
-import { StudioConnectCard } from './StudioConnectCard.js';
-import { StudioCreatorAgentKeyPanel } from './StudioCreatorAgentKeyPanel.js';
-import { StudioDetailsBuildProgress } from './StudioDetailsBuildProgress.js';
-import { StudioDetailsMedia } from './StudioDetailsMedia.js';
-import { ContributionsSetting } from '../../ContributionsSetting.js';
-import { ProposalReviewPanel } from '../review/ProposalReviewPanel.js';
-import { StudioOAuthClientsPanel } from './StudioOAuthClientsPanel.js';
-import { StudioWorkspaceCheckoutPanel } from './StudioWorkspaceCheckoutPanel.js';
+import { DetailsPanel, type DetailsPaneId } from './StudioDetailsPanel.js';
+import { DraftShareControl } from './DraftShareControl.js';
+import { StudioShelfControls, StudioShelfList } from './StudioShelf.js';
+import { defaultTabFor, resolveTab, studioAddress, tabAvailable } from './studioTabs.js';
+import { healthFor } from './studioHealth.js';
 import { SubmissionStatusView } from './SubmissionStatusView.js';
 import {
-  approveSuggestion,
-  dismissSuggestion,
   fetchStudioGames,
   fetchStudioHealth,
   fetchStudioScorecards,
-  fetchStudioSuggestions,
-  fetchGameAutonomy,
-  setGameAutonomy,
-  setDraftShared,
   type StudioGame,
   type StudioScorecard,
-  type StudioSuggestion,
-  type AutonomyMode,
 } from '../../studioApi.js';
 import type { EditorContentPush, EditorControllerState } from '../../editorBridge.js';
 import './studio-connect.css';
@@ -97,8 +83,6 @@ import './studio-stats.css';
  * whichever surface absorbed them.
  */
 
-const WINDOWS = [1, 7, 30];
-
 // Keep in step with the same breakpoint in styles.css.
 const SHEET_MAX_WIDTH = 1099;
 const SHEET_QUERY = `(max-width: ${SHEET_MAX_WIDTH}px)`;
@@ -125,69 +109,6 @@ type CreatorStudioViewProps = {
   /** Loads a failed/abandoned concept back into the home hero prompt. */
   onRetryConcept?: (concept: string) => void;
 };
-
-/**
- * The thread, always — for a game being built and for one that is live.
- *
- * Both used to land somewhere else: a build opened on its status page, a published game
- * on a facts panel. But the question a creator arrives with is the same either way —
- * what has happened, and how do I ask for the next thing — and that is one surface.
- */
-function defaultTabFor(): StudioTab {
-  return 'thread';
-}
-
-/**
- * Which surfaces exist for this game. All three, for every game: the thread reads the
- * build's history and takes the next message whatever state it is in, the details panel
- * always has facts to show, and playtest always has something to play or a reason it
- * does not yet.
- */
-function tabAvailable(game: StudioGame, tab: StudioTab): boolean {
-  // The Edit surface exists only for games whose latest build ships an editor
-  // definition. Every other game keeps exactly the three surfaces it had — an
-  // /edit URL for one of those resolves and falls back to the thread.
-  if (tab === 'edit') return Boolean(game.editable && game.slug);
-  // The Code surface (CE-06): every owned game with a slug, once the kill switch is
-  // on — no manifest condition, unlike Edit, per the owner decision (all creators
-  // from M1). `game.slug` still gates it: there is nothing to read or write before
-  // the game has one.
-  if (tab === 'code') return Boolean(game.codeSurface && game.slug);
-  return true;
-}
-
-function resolveTab(game: StudioGame, requested?: StudioTab): StudioTab {
-  if (requested && tabAvailable(game, requested)) return requested;
-  return defaultTabFor();
-}
-
-/**
- * How this game is addressed in the URL.
- *
- * The slug once it has one, which is now from the moment the game was submitted. The
- * capability token is the fallback for games created before slugs were assigned up
- * front — and it is a fallback rather than the rule because a token in the URL bar is a
- * grant sitting in browser history, in screenshots, and in anything the creator pastes.
- */
-function studioAddress(game: StudioGame): string {
-  return game.slug ?? game.token;
-}
-
-function formatSeconds(seconds: number): string {
-  if (seconds < 60) return `${Math.round(seconds)}s`;
-  const minutes = Math.floor(seconds / 60);
-  const rest = Math.round(seconds % 60);
-  return rest === 0 ? `${minutes}m` : `${minutes}m ${rest}s`;
-}
-
-function percent(rate: number): string {
-  return `${Math.round(rate * 100)}%`;
-}
-
-function healthFor(game: StudioGame, rows: GameHealth[]): GameHealth | null {
-  if (!game.slug) return null;
-  return rows.find((row) => row.slug === game.slug) ?? null;
-}
 
 export function CreatorStudioView({
   selectedGame,
@@ -1269,979 +1190,3 @@ export function CreatorStudioView({
     </StudioCreatorProfileProvider>
   );
 }
-
-function StudioShelfControls({
-  searchInputId,
-  searchRef,
-  query,
-  filter,
-  showTools,
-  buildingCount,
-  liveCount,
-  totalCount,
-  onQueryChange,
-  onFilterChange,
-}: {
-  searchInputId: string;
-  searchRef?: RefObject<HTMLInputElement>;
-  query: string;
-  filter: StudioShelfFilter;
-  showTools: boolean;
-  buildingCount: number;
-  liveCount: number;
-  totalCount: number;
-  onQueryChange: (value: string) => void;
-  onFilterChange: (value: StudioShelfFilter) => void;
-}) {
-  const { t } = useTranslation();
-  if (!showTools) return null;
-
-  return (
-    <div className="studio-shelf-tools">
-      <label className="studio-shelf-search" htmlFor={searchInputId}>
-        <PixelIcon name="search" size={12} />
-        <span className="studio-sr-only">{t('studioPanel.shelf.searchLabel')}</span>
-        <input
-          id={searchInputId}
-          ref={searchRef}
-          type="search"
-          value={query}
-          onChange={(event) => onQueryChange(event.target.value)}
-          placeholder={t('studioPanel.shelf.searchPlaceholder')}
-          autoComplete="off"
-        />
-      </label>
-      <div className="studio-shelf-filters" role="group" aria-label={t('studioPanel.shelf.filterAria')}>
-        {(
-          [
-            ['all', t('studioPanel.shelf.filters.all'), totalCount],
-            ['building', t('studioPanel.shelf.filters.building'), buildingCount],
-            ['live', t('studioPanel.shelf.filters.live'), liveCount],
-          ] as const
-        ).map(([id, label, count]) => (
-          <button
-            key={id}
-            type="button"
-            className={`studio-shelf-filter${filter === id ? ' is-active' : ''}`}
-            aria-pressed={filter === id}
-            onClick={() => onFilterChange(id)}
-          >
-            {label}
-            <span className="studio-shelf-filter-count">{count}</span>
-          </button>
-        ))}
-      </div>
-    </div>
-  );
-}
-
-function StudioShelfList({
-  games,
-  selected,
-  locale,
-  emptyLabel,
-  onSelect,
-}: {
-  games: StudioShelfGame[];
-  selected: string | null;
-  locale: string;
-  emptyLabel: string;
-  onSelect: (token: string) => void;
-}) {
-  const { t } = useTranslation();
-
-  if (games.length === 0) {
-    return <p className="studio-shelf-empty">{emptyLabel}</p>;
-  }
-
-  return (
-    <ul className="studio-shelf-list">
-      {games.map((game) => {
-        const active = game.token === selected;
-        const status = game.lastKnownStatus;
-        const building = Boolean(status && STUDIO_LIVE_STATUSES.has(status));
-        // Building wins the dot: a revise tip on a live game is still "moving".
-        const live = !building && isStudioGameShelfLive(game);
-        return (
-          <li key={game.token}>
-            <button
-              type="button"
-              className={`studio-shelf-item${active ? ' is-active' : ''}${building ? ' is-live' : ''}${live ? ' is-published' : ''}`}
-              onClick={() => onSelect(game.token)}
-              aria-current={active ? 'true' : undefined}
-              title={game.title}
-            >
-              <span
-                className={`studio-shelf-mark${building ? ' is-live' : ''}${live ? ' is-published' : ''}`}
-                aria-hidden="true"
-              >
-                {studioGameInitials(game.title)}
-              </span>
-              <span className="studio-shelf-title">{game.title}</span>
-              <span className="studio-sr-only">
-                {status ? t(`statusView.states.${status}.label`) : t('myGames.checking')} ·{' '}
-                {formatRelativeTime(Date.parse(game.createdAt), locale)}
-              </span>
-            </button>
-          </li>
-        );
-      })}
-    </ul>
-  );
-}
-
-/** Cursor-style Details strip — one pane at a time, chosen by icon. */
-type DetailsPaneId = 'overview' | 'connect' | 'build' | 'media' | 'workspace' | 'keys' | 'stats';
-
-type DetailsPaneDef = {
-  id: DetailsPaneId;
-  icon: PixelIconName;
-  labelKey: string;
-};
-
-/**
- * The things beside the thread: when the game was made, how it is doing, who can play
- * it, and how to stop it. Everything here is *about* the game; the thread is where it
- * is talked to. Layout mirrors Cursor’s secondary icon rail — icons pick one pane.
- */
-function DetailsPanel({
-  game,
-  mediaToken,
-  health,
-  days,
-  healthDays,
-  truncated,
-  scorecard,
-  pane,
-  onPaneChange,
-  onClose,
-  onDaysChange,
-  onOpenPlaytest,
-  onSelectPreviewVersion,
-  activePreviewVersion,
-  onReverted,
-  onSwitchToPlatform,
-  onPlay,
-  onDraftSharedChange,
-  onRemoved,
-}: {
-  game: StudioShelfGame;
-  /** Token for Media (and the active build round). Differs from `game.token` during handoff. */
-  mediaToken?: string;
-  health: GameHealth | null;
-  days: number;
-  healthDays: string[];
-  truncated: boolean;
-  scorecard: StudioScorecard | null;
-  pane: DetailsPaneId;
-  onPaneChange: (pane: DetailsPaneId) => void;
-  onClose: () => void;
-  onDaysChange: (days: number) => void;
-  onOpenPlaytest: () => void;
-  onSelectPreviewVersion?: (version: string | null) => void;
-  activePreviewVersion?: string | null;
-  onReverted?: (result: { version: string; token?: string; roundOpened?: number }) => void;
-  onSwitchToPlatform: () => Promise<void>;
-  onPlay: () => void;
-  onDraftSharedChange: (shared: boolean) => void;
-  onRemoved: (token: string) => void | Promise<void>;
-}) {
-  const { t, i18n } = useTranslation();
-  // This *job* is published — composer/playtest routing. Distinct from catalog-live below.
-  const publishedJob = isStudioGamePublished(game);
-  const catalogLive = isStudioGameShelfLive(game);
-  const publishedAt = game.publishedAt ?? game.livePublishedAt;
-  const [abandonArmed, setAbandonArmed] = useState(false);
-  const [abandoning, setAbandoning] = useState(false);
-  const [deleteArmed, setDeleteArmed] = useState(false);
-  const [deleting, setDeleting] = useState(false);
-  const shotToken = mediaToken ?? game.token;
-
-  async function handleAbandon() {
-    if (!abandonArmed) {
-      setAbandonArmed(true);
-      return;
-    }
-    setAbandoning(true);
-    try {
-      await abandonSubmission(game.token);
-      await onRemoved(game.token);
-    } catch {
-      setAbandoning(false);
-      setAbandonArmed(false);
-    }
-  }
-
-  async function handleDelete() {
-    if (!deleteArmed) {
-      setDeleteArmed(true);
-      return;
-    }
-    setDeleting(true);
-    try {
-      await deleteGame(game.token);
-      await onRemoved(game.token);
-    } catch {
-      setDeleting(false);
-      setDeleteArmed(false);
-    }
-  }
-
-  const showConnect = game.lastKnownStatus !== 'abandoned' && game.lastKnownStatus !== 'published';
-  // A published game's own status hides Build — but an improvement handoff runs a live
-  // round under `mediaToken`, distinct from `game.token`, while `game` still reads published.
-  const inHandoffRound = Boolean(mediaToken && mediaToken !== game.token);
-  const showProgress = showConnect || inHandoffRound;
-  const panes: DetailsPaneDef[] = [
-    { id: 'overview', icon: 'eye', labelKey: 'studioPanel.rail.overview' },
-    ...(showConnect ? [{ id: 'connect' as const, icon: 'signal' as const, labelKey: 'studioPanel.rail.connect' }] : []),
-    ...(showProgress ? [{ id: 'build' as const, icon: 'wrench' as const, labelKey: 'studioPanel.rail.build' }] : []),
-    { id: 'media', icon: 'image', labelKey: 'studioPanel.rail.media' },
-    // No slug means no sources yet, and the route would refuse it.
-    ...(game.slug && game.lastKnownStatus !== 'abandoned'
-      ? [{ id: 'workspace' as const, icon: 'download' as const, labelKey: 'studioPanel.rail.workspace' }]
-      : []),
-    { id: 'keys', icon: 'lock', labelKey: 'studioPanel.rail.credentials' },
-    ...(catalogLive ? [{ id: 'stats' as const, icon: 'star' as const, labelKey: 'studioPanel.rail.stats' }] : []),
-  ];
-
-  // If the open pane disappeared (e.g. published while on Connect), fall back.
-  const activePane = panes.some((entry) => entry.id === pane) ? pane : 'overview';
-  const activeLabel = t(panes.find((entry) => entry.id === activePane)?.labelKey ?? 'studioPanel.tabs.details');
-
-  return (
-    <div className="studio-rail-shell" data-testid="studio-rail-shell">
-      <div className="studio-rail-head">
-        <h3>{activeLabel}</h3>
-        <button type="button" className="modal-close-btn" onClick={onClose} aria-label={t('studioPanel.close')}>
-          <PixelIcon name="close" size={14} />
-        </button>
-      </div>
-
-      <div className="studio-rail-pane studio-overview" data-testid={`studio-rail-pane-${activePane}`}>
-        {activePane === 'overview' ? (
-          <>
-            <section className="studio-rail-section" aria-label={t('studioPanel.overview.status')}>
-              <ul className="funnel-stats studio-facts">
-                <li>
-                  <span className="funnel-stat-value">
-                    {formatRelativeTime(Date.parse(game.createdAt), i18n.language)}
-                  </span>
-                  <span className="funnel-stat-label">{t('studioPanel.overview.created')}</span>
-                </li>
-                {publishedAt ? (
-                  <li>
-                    <span className="funnel-stat-value">
-                      {formatRelativeTime(Date.parse(publishedAt), i18n.language)}
-                    </span>
-                    <span className="funnel-stat-label">{t('studioPanel.overview.published')}</span>
-                  </li>
-                ) : null}
-                {health ? (
-                  <li>
-                    <span className="funnel-stat-value">
-                      {health.sessions}
-                      <span className="studio-fact-suffix">
-                        · {formatSeconds(health.totalPlaySeconds)} {t('studioPanel.overview.play')}
-                      </span>
-                    </span>
-                    <span className="funnel-stat-label">{t('studioPanel.overview.sessions')}</span>
-                  </li>
-                ) : null}
-              </ul>
-
-              <div className="studio-actions">
-                {catalogLive && game.slug ? (
-                  <button type="button" className="primary-btn" onClick={onPlay}>
-                    <PixelIcon name="play" size={12} /> {t('myGames.play')}
-                  </button>
-                ) : null}
-                <button type="button" className="secondary-btn" onClick={onOpenPlaytest}>
-                  <PixelIcon name="play" size={12} /> {t('studioPanel.overview.playtest')}
-                </button>
-                {!publishedJob && game.lastKnownStatus !== 'abandoned' ? (
-                  <div className="studio-abandon-block">
-                    {abandonArmed && !catalogLive ? (
-                      <p className="studio-abandon-hint">{t('studioPanel.overview.abandonHintRemove')}</p>
-                    ) : null}
-                    <button
-                      type="button"
-                      className={`status-abandon${abandonArmed ? ' is-danger' : ''}`}
-                      onClick={() => void handleAbandon()}
-                      disabled={abandoning}
-                    >
-                      {abandonArmed
-                        ? t(
-                            catalogLive
-                              ? 'studioPanel.overview.abandonConfirmKeepLive'
-                              : 'studioPanel.overview.abandonConfirmRemove',
-                          )
-                        : t(catalogLive ? 'studioPanel.overview.abandon' : 'studioPanel.overview.abandonRemove')}
-                    </button>
-                  </div>
-                ) : null}
-                {catalogLive && game.lastKnownStatus !== 'abandoned' ? (
-                  <div className="studio-delete-block">
-                    {deleteArmed ? <p className="studio-delete-hint">{t('studioPanel.overview.deleteHint')}</p> : null}
-                    <button
-                      type="button"
-                      className={`status-delete${deleteArmed ? ' is-danger' : ''}`}
-                      onClick={() => void handleDelete()}
-                      disabled={deleting}
-                    >
-                      {deleteArmed ? t('studioPanel.overview.deleteConfirm') : t('studioPanel.overview.delete')}
-                    </button>
-                  </div>
-                ) : null}
-              </div>
-            </section>
-
-            {game.slug && game.lastKnownStatus !== 'abandoned' ? (
-              <section
-                className="studio-rail-section"
-                aria-label={t(catalogLive ? 'studioPanel.share.liveTitle' : 'studioPanel.share.title')}
-              >
-                <DraftShareControl game={game} live={catalogLive} onSharedChange={onDraftSharedChange} />
-              </section>
-            ) : null}
-          </>
-        ) : null}
-
-        {/*
-         * Contributions live in the overview pane rather than getting a pane of their
-         * own: for the great majority of games the answer is "off" and stays off, and a
-         * tab per setting is how a rail becomes a menu. The review cards sit directly
-         * under the switch that produced them, so turning it on and seeing what arrives
-         * is one place rather than two.
-         *
-         * Only for a published game with a slug: there is nothing to propose against a
-         * draft, and offering the switch there would promise a door that does not exist.
-         */}
-        {activePane === 'overview' && game.slug && catalogLive ? (
-          <div className="studio-rail-contributions">
-            <ContributionsSetting slug={game.slug} />
-            <ProposalReviewPanel scope="mine" slug={game.slug} />
-          </div>
-        ) : null}
-
-        {activePane === 'connect' ? (
-          showConnect ? (
-            <StudioConnectCard
-              token={game.token}
-              collapsible={false}
-              hideIfUnavailable
-              unavailableLabel={t('studioPanel.rail.connectEmpty')}
-              density="panel"
-              onSwitchToPlatform={onSwitchToPlatform}
-            />
-          ) : (
-            <p className="studio-rail-empty">{t('studioPanel.rail.connectEmpty')}</p>
-          )
-        ) : null}
-
-        {activePane === 'build' ? (
-          showProgress ? (
-            <StudioDetailsBuildProgress
-              token={shotToken}
-              emptyLabel={t('studioPanel.rail.buildEmpty')}
-              onSelectPreviewVersion={onSelectPreviewVersion}
-              activePreviewVersion={activePreviewVersion}
-              onReverted={onReverted}
-            />
-          ) : (
-            <p className="studio-rail-empty">{t('studioPanel.rail.buildEmpty')}</p>
-          )
-        ) : null}
-
-        {activePane === 'media' ? (
-          <StudioDetailsMedia token={shotToken} emptyLabel={t('studioPanel.rail.mediaEmpty')} />
-        ) : null}
-
-        {activePane === 'keys' ? (
-          <div className="studio-rail-credentials-body">
-            <p className="studio-rail-credentials-hint">{t('studioPanel.rail.credentialsHint')}</p>
-            <StudioCreatorAgentKeyPanel />
-            <StudioOAuthClientsPanel />
-          </div>
-        ) : null}
-
-        {/* Own pane: behind the keys icon it was undiscoverable. */}
-        {activePane === 'workspace' ? <StudioWorkspaceCheckoutPanel slug={game.slug as string} /> : null}
-
-        {activePane === 'stats' && catalogLive ? (
-          <StatsSection
-            game={game}
-            health={health}
-            days={days}
-            healthDays={healthDays}
-            truncated={truncated}
-            scorecard={scorecard}
-            onDaysChange={onDaysChange}
-          />
-        ) : null}
-      </div>
-
-      <nav className="studio-rail-icons" aria-label={t('studioPanel.tabs.details')}>
-        {panes.map((entry) => (
-          <button
-            key={entry.id}
-            type="button"
-            className={`studio-rail-icon${activePane === entry.id ? ' is-active' : ''}`}
-            aria-pressed={activePane === entry.id}
-            aria-label={t(entry.labelKey)}
-            title={t(entry.labelKey)}
-            data-testid={`studio-rail-icon-${entry.id}`}
-            onClick={() => onPaneChange(entry.id)}
-          >
-            <PixelIcon name={entry.icon} size={14} />
-          </button>
-        ))}
-      </nav>
-    </div>
-  );
-}
-
-/**
- * Whether anyone else can play this game before it is published.
- *
- * Off until the creator turns it on, and the game is in no catalog and no rail either
- * way — the link is the only way in, which is what makes one switch enough. The link
- * shown is the game's ordinary permalink, the same one it will keep once it is live,
- * so there is nothing to re-share when that happens.
- *
- * Once the game *is* live in the catalog there is no switch left to show — the link is
- * public by definition — but the creator still wants to hand it out, so `live` drops the
- * toggle and keeps the permalink and its copy button.
- */
-function DraftShareControl({
-  game,
-  compact = false,
-  live = false,
-  onSharedChange,
-}: {
-  game: StudioGame;
-  /** Drop the card chrome when nested in the header popover. */
-  compact?: boolean;
-  /** Game is live in the catalog: permalink only, no draft switch. */
-  live?: boolean;
-  onSharedChange?: (shared: boolean) => void;
-}) {
-  const { t } = useTranslation();
-  const [shared, setShared] = useState(Boolean(game.draftShared));
-  const [busy, setBusy] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const [copied, setCopied] = useState(false);
-  // Header popover unmounts this control on Escape / game switch; ignore the
-  // in-flight toggle result so we do not setState after unmount.
-  const mountedRef = useRef(true);
-
-  useEffect(() => {
-    mountedRef.current = true;
-    return () => {
-      mountedRef.current = false;
-    };
-  }, []);
-
-  useEffect(() => {
-    setShared(Boolean(game.draftShared));
-  }, [game.draftShared, game.token]);
-
-  const url = game.slug ? new URL(playPath(game.slug), window.location.href).toString() : '';
-
-  async function toggle() {
-    const next = !shared;
-    setBusy(true);
-    setError(null);
-    // Optimistic, and reverted on failure: the switch is the whole control, so leaving
-    // it in the old position while the request runs reads as a dead button.
-    setShared(next);
-    try {
-      await setDraftShared(game.token, next);
-      if (!mountedRef.current) return;
-      onSharedChange?.(next);
-    } catch {
-      if (!mountedRef.current) return;
-      setShared(!next);
-      setError(t('studioPanel.share.error'));
-    } finally {
-      if (mountedRef.current) setBusy(false);
-    }
-  }
-
-  async function copy() {
-    try {
-      await navigator.clipboard.writeText(url);
-      if (!mountedRef.current) return;
-      setCopied(true);
-      window.setTimeout(() => {
-        if (mountedRef.current) setCopied(false);
-      }, 2000);
-    } catch {
-      // No clipboard permission or no clipboard API — the link is on screen to select
-      // by hand, so this needs no error state.
-    }
-  }
-
-  return (
-    <div className={`studio-share${compact ? ' is-compact' : ''}${live ? ' is-live' : ''}`}>
-      <div className="studio-share-head">
-        <h3 className="studio-share-title">{t(live ? 'studioPanel.share.liveTitle' : 'studioPanel.share.title')}</h3>
-        {live ? null : (
-          <button
-            type="button"
-            role="switch"
-            aria-checked={shared}
-            className={`studio-share-toggle${shared ? ' is-on' : ''}`}
-            onClick={() => void toggle()}
-            disabled={busy}
-          >
-            <span className="studio-share-toggle-track" aria-hidden="true" />
-            {shared ? t('studioPanel.share.on') : t('studioPanel.share.off')}
-          </button>
-        )}
-      </div>
-      <p className="studio-share-hint">
-        {t(live ? 'studioPanel.share.liveHint' : shared ? 'studioPanel.share.hintOn' : 'studioPanel.share.hintOff')}
-      </p>
-      {live || shared ? (
-        <p className="status-note status-share">
-          <a className="inline-link" href={url}>
-            {url}
-          </a>
-          <button type="button" className="status-share-copy" onClick={() => void copy()}>
-            <PixelIcon name={copied ? 'check' : 'globe'} size={12} />{' '}
-            {copied ? t('statusView.shareCopied') : t('statusView.shareCopy')}
-          </button>
-        </p>
-      ) : null}
-      {error ? <p className="error">{error}</p> : null}
-    </div>
-  );
-}
-
-function StatsSection({
-  game,
-  health,
-  days,
-  healthDays,
-  truncated,
-  scorecard,
-  onDaysChange,
-}: {
-  game: StudioGame;
-  health: GameHealth | null;
-  days: number;
-  healthDays: string[];
-  truncated: boolean;
-  scorecard: StudioScorecard | null;
-  onDaysChange: (days: number) => void;
-}) {
-  const { t } = useTranslation();
-
-  if (!game.slug) {
-    return <p className="studio-empty">{t('studioPanel.stats.noSlug')}</p>;
-  }
-
-  return (
-    <div className="studio-stats">
-      <div className="health-windows">
-        {WINDOWS.map((window) => (
-          <button
-            key={window}
-            type="button"
-            className={window === days ? 'health-window is-active' : 'health-window'}
-            onClick={() => onDaysChange(window)}
-          >
-            {window}d
-          </button>
-        ))}
-      </div>
-
-      {healthDays.length > 0 ? (
-        <p className="studio-stats-range">
-          {t('studioPanel.stats.range', { from: healthDays[healthDays.length - 1], to: healthDays[0] })}
-        </p>
-      ) : null}
-      {truncated ? <p className="health-note">{t('studioPanel.stats.truncated')}</p> : null}
-
-      {!health || health.sessions === 0 ? (
-        <p className="studio-empty">{t('studioPanel.stats.empty')}</p>
-      ) : (
-        <ul className="funnel-stats">
-          <li>
-            <span className="funnel-stat-value">{health.sessions}</span>
-            <span className="funnel-stat-label">{t('studioPanel.stats.sessions')}</span>
-          </li>
-          <li>
-            <span className="funnel-stat-value">
-              {health.bounces} ({percent(health.sessions === 0 ? 0 : health.bounces / health.sessions)})
-            </span>
-            <span className="funnel-stat-label">{t('studioPanel.stats.bounces')}</span>
-          </li>
-          <li>
-            <span className="funnel-stat-value">{formatSeconds(health.medianPlaySeconds)}</span>
-            <span className="funnel-stat-label">{t('studioPanel.stats.medianPlay')}</span>
-          </li>
-          <li>
-            <span className="funnel-stat-value">{formatSeconds(health.totalPlaySeconds)}</span>
-            <span className="funnel-stat-label">{t('studioPanel.stats.totalPlay')}</span>
-          </li>
-          <li>
-            <span className="funnel-stat-value">{health.errors}</span>
-            <span className="funnel-stat-label">{t('studioPanel.stats.errors')}</span>
-          </li>
-          <li>
-            <span className="funnel-stat-value">{percent(health.stallRate)}</span>
-            <span className="funnel-stat-label">{t('studioPanel.stats.stallRate')}</span>
-          </li>
-          <li>
-            <span className="funnel-stat-value">{health.medianFps === null ? '—' : Math.round(health.medianFps)}</span>
-            <span className="funnel-stat-label">{t('studioPanel.stats.medianFps')}</span>
-          </li>
-        </ul>
-      )}
-
-      <PlayerReactions scorecard={scorecard} />
-
-      <SuggestedImprovements slug={game.slug} />
-
-      <AutonomySetting slug={game.slug} />
-
-      {health && health.errorSamples.length > 0 ? (
-        <div className="studio-error-samples">
-          <h3 className="health-section-title">{t('studioPanel.stats.errorSamples')}</h3>
-          <ul>
-            {health.errorSamples.map((sample) => (
-              <li key={sample.message}>
-                <code>{sample.message}</code>
-                <span className="health-error-count">×{sample.count}</span>
-              </li>
-            ))}
-          </ul>
-        </div>
-      ) : null}
-    </div>
-  );
-}
-
-/**
- * Votes and what players wrote — the third question the stats tab has to answer.
- *
- * Separated visually from the numbers above because it is measured over a different
- * window: the health block recomputes over the window the creator picked, while these come
- * from the nightly scorecard's fixed roll. Two windows on one screen is fine; two windows
- * that look like one is not, so the period is stated.
- *
- * Themes are player-written text summarized by a model. React escapes them, which is what
- * makes showing them safe; the label is what stops them being mistaken for something the
- * system is asserting.
- */
-function PlayerReactions({ scorecard }: { scorecard: StudioScorecard | null }) {
-  const { t } = useTranslation();
-
-  // Absent, not zero: no scorecard means this game has not been rolled up yet, which is
-  // not the same as a game measured and found to have no reactions.
-  if (!scorecard) return null;
-
-  const themes = scorecard.untrustedThemes;
-  const nothingYet = scorecard.votes.up === 0 && scorecard.votes.down === 0 && scorecard.feedbackCount === 0;
-
-  return (
-    <div className="studio-reactions">
-      <h3 className="health-section-title">{t('studioPanel.stats.reactions')}</h3>
-      <p className="studio-stats-range">{t('studioPanel.stats.reactionsWindow', { days: scorecard.windowDays })}</p>
-
-      {nothingYet ? (
-        <p className="studio-empty">{t('studioPanel.stats.reactionsEmpty')}</p>
-      ) : (
-        <ul className="funnel-stats">
-          <li>
-            <span className="funnel-stat-value">
-              {scorecard.votes.up}↑ {scorecard.votes.down}↓
-            </span>
-            <span className="funnel-stat-label">{t('studioPanel.stats.votes')}</span>
-          </li>
-          <li>
-            <span className="funnel-stat-value">{scorecard.feedbackCount}</span>
-            <span className="funnel-stat-label">{t('studioPanel.stats.notes')}</span>
-          </li>
-        </ul>
-      )}
-
-      {themes.length > 0 ? (
-        <div className="studio-themes">
-          <h4 className="studio-themes-title">{t('studioPanel.stats.themes')}</h4>
-          <p className="health-note">{t('studioPanel.stats.themesNote')}</p>
-          <ul className="studio-theme-list">
-            {themes.map((entry) => (
-              <li key={entry.theme}>
-                {entry.theme} <span className="health-error-count">×{entry.count}</span>
-              </li>
-            ))}
-          </ul>
-        </div>
-      ) : null}
-    </div>
-  );
-}
-
-/**
- * The suggestion inbox for one game (docs/improvement-loop-plan.md IL-3).
- *
- * Cards read insight → evidence → decide. Two things are deliberate here.
- *
- * The **evidence** the platform measured is stated plainly, while the game's and players'
- * own words sit in a separate block labelled as such. React escapes both, so neither is
- * a markup risk; the separation is about not letting a string somebody else chose read
- * as though this platform were asserting it.
- *
- * **Approving can succeed without an implementer.** The API records the decision and
- * reports `no-implementer` when the coding agent could not be reached, so this renders
- * that as a real outcome with a retry rather than as a failure — the creator's click
- * counted either way.
- */
-function SuggestedImprovements({ slug }: { slug: string | undefined }) {
-  const { t } = useTranslation();
-  const [suggestions, setSuggestions] = useState<StudioSuggestion[] | null>(null);
-  const [busyId, setBusyId] = useState<string | null>(null);
-  const [error, setError] = useState<string | null>(null);
-  const [dismissing, setDismissing] = useState<string | null>(null);
-
-  useEffect(() => {
-    let cancelled = false;
-    fetchStudioSuggestions()
-      .then((rows) => {
-        if (!cancelled) setSuggestions(rows);
-      })
-      // A queue that fails to load must not take the stats page down with it.
-      .catch(() => {
-        if (!cancelled) setSuggestions([]);
-      });
-    return () => {
-      cancelled = true;
-    };
-  }, []);
-
-  const mine = useMemo(
-    () => (suggestions ?? []).filter((entry) => entry.slug === slug && entry.status === 'proposed'),
-    [suggestions, slug],
-  );
-
-  const replace = (updated: StudioSuggestion) =>
-    setSuggestions((rows) => (rows ?? []).map((row) => (row.id === updated.id ? updated : row)));
-
-  const decided = (suggestions ?? []).filter(
-    (entry) => entry.slug === slug && (entry.status === 'dispatched' || entry.status === 'no-implementer'),
-  );
-
-  async function act(id: string, run: () => Promise<StudioSuggestion>) {
-    setBusyId(id);
-    setError(null);
-    try {
-      replace(await run());
-      setDismissing(null);
-    } catch (caught) {
-      const status = (caught as { status?: number }).status;
-      setError(status === 429 ? t('studioPanel.suggestions.quota') : t('studioPanel.suggestions.failed'));
-    } finally {
-      setBusyId(null);
-    }
-  }
-
-  if (suggestions === null || (mine.length === 0 && decided.length === 0)) return null;
-
-  return (
-    <div className="studio-suggestions">
-      <h3 className="health-section-title">{t('studioPanel.suggestions.title')}</h3>
-      <p className="health-note">{t('studioPanel.suggestions.note')}</p>
-      {error ? <p className="studio-error">{error}</p> : null}
-
-      {decided.map((entry) => (
-        <p key={entry.id} className="studio-suggestion-outcome">
-          {entry.status === 'dispatched'
-            ? t('studioPanel.suggestions.filed')
-            : t('studioPanel.suggestions.noImplementer')}
-        </p>
-      ))}
-
-      {mine.map((entry) => (
-        <article key={entry.id} className="studio-suggestion">
-          <h4 className="studio-suggestion-class">{classLabel(entry.class, t)}</h4>
-
-          <ul className="studio-suggestion-evidence">
-            {entry.evidence.map((item) => (
-              <li key={item.finding}>{item.finding}</li>
-            ))}
-          </ul>
-
-          <SuggestionContext context={entry.untrustedContext} />
-
-          {dismissing === entry.id ? (
-            <div className="studio-suggestion-reasons">
-              <p>{t('studioPanel.suggestions.dismissReason')}</p>
-              {DISMISS_REASON_KEYS.map(([reason, key]) => (
-                <button
-                  key={reason}
-                  type="button"
-                  className="studio-suggestion-reason"
-                  disabled={busyId === entry.id}
-                  onClick={() => act(entry.id, () => dismissSuggestion(entry.id, reason))}
-                >
-                  {t(key)}
-                </button>
-              ))}
-            </div>
-          ) : (
-            <div className="studio-suggestion-actions">
-              <button
-                type="button"
-                className="studio-suggestion-approve"
-                disabled={busyId === entry.id}
-                onClick={() => act(entry.id, () => approveSuggestion(entry.id))}
-                title={t('studioPanel.suggestions.approveHint')}
-              >
-                {t('studioPanel.suggestions.approve')}
-              </button>
-              <button
-                type="button"
-                className="studio-suggestion-dismiss"
-                disabled={busyId === entry.id}
-                onClick={() => setDismissing(entry.id)}
-              >
-                {t('studioPanel.suggestions.dismiss')}
-              </button>
-            </div>
-          )}
-        </article>
-      ))}
-    </div>
-  );
-}
-
-/** The fixed dismissal vocabulary the API accepts, paired with its translation key. */
-const DISMISS_REASON_KEYS: Array<[DismissReason, string]> = [
-  ['intentional', 'studioPanel.suggestions.reasonIntentional'],
-  ['not-a-problem', 'studioPanel.suggestions.reasonNotAProblem'],
-  ['wont-fix', 'studioPanel.suggestions.reasonWontFix'],
-  ['not-now', 'studioPanel.suggestions.reasonNotNow'],
-  ['bad-evidence', 'studioPanel.suggestions.reasonBadEvidence'],
-];
-
-function classLabel(suggestionClass: string, t: (key: string) => string): string {
-  if (suggestionClass === 'defect') return t('studioPanel.suggestions.classDefect');
-  if (suggestionClass === 'friction') return t('studioPanel.suggestions.classFriction');
-  if (suggestionClass === 'design-change') return t('studioPanel.suggestions.classDesignChange');
-  if (suggestionClass === 'editorial') return t('studioPanel.suggestions.classEditorial');
-  return suggestionClass;
-}
-
-/** Game- and player-authored strings, kept visually separate from what we measured. */
-function SuggestionContext({ context }: { context: StudioSuggestion['untrustedContext'] }) {
-  const { t } = useTranslation();
-  const samples = context?.errorSamples ?? [];
-  const themes = context?.feedbackThemes ?? [];
-  if (!samples.length && !themes.length) return null;
-
-  return (
-    <div className="studio-suggestion-context">
-      <h5 className="studio-themes-title">{t('studioPanel.suggestions.context')}</h5>
-      <p className="health-note">{t('studioPanel.suggestions.contextNote')}</p>
-      <ul className="studio-theme-list">
-        {samples.map((sample) => (
-          <li key={sample.message}>
-            {sample.message} <span className="health-error-count">×{sample.count}</span>
-          </li>
-        ))}
-        {themes.map((theme) => (
-          <li key={theme.theme}>
-            {theme.theme} <span className="health-error-count">×{theme.count}</span>
-          </li>
-        ))}
-      </ul>
-    </div>
-  );
-}
-
-/**
- * What the platform may do to this game without asking (IL-4).
- *
- * Framed as permission rather than as a feature toggle, and it says the reassuring part
- * out loud: nothing reaches the site without the creator's review whatever they pick.
- * That is not marketing — `publishing` is reachable only from `ready_for_review` in the
- * job state machine, so it is a property of the system rather than a promise about it.
- */
-function AutonomySetting({ slug }: { slug: string | undefined }) {
-  const { t } = useTranslation();
-  const [mode, setMode] = useState<AutonomyMode | null>(null);
-  const [state, setState] = useState<'idle' | 'saving' | 'saved' | 'error'>('idle');
-
-  useEffect(() => {
-    if (!slug) return;
-    let cancelled = false;
-    fetchGameAutonomy(slug)
-      .then((value) => {
-        if (!cancelled) setMode(value);
-      })
-      // A game the creator does not own, or a deployment without this route, simply has
-      // no control to show — it must not break the stats page around it.
-      .catch(() => {
-        if (!cancelled) setMode(null);
-      });
-    return () => {
-      cancelled = true;
-    };
-  }, [slug]);
-
-  if (!slug || mode === null) return null;
-
-  async function choose(next: AutonomyMode) {
-    if (!slug) return;
-    const previous = mode;
-    setMode(next);
-    setState('saving');
-    try {
-      setMode(await setGameAutonomy(slug, next));
-      setState('saved');
-    } catch {
-      // Put it back rather than leave the control showing a setting that is not stored.
-      setMode(previous);
-      setState('error');
-    }
-  }
-
-  return (
-    <div className="studio-autonomy">
-      <h3 className="health-section-title">{t('studioPanel.autonomy.title')}</h3>
-      <p className="health-note">{t('studioPanel.autonomy.note')}</p>
-      <ul className="studio-autonomy-options">
-        {AUTONOMY_CHOICES.map(([value, labelKey, hintKey]) => (
-          <li key={value}>
-            <label className={value === mode ? 'studio-autonomy-option is-active' : 'studio-autonomy-option'}>
-              <input
-                type="radio"
-                name={`autonomy-${slug}`}
-                checked={value === mode}
-                disabled={state === 'saving'}
-                onChange={() => choose(value)}
-              />
-              <span>
-                <strong>{t(labelKey)}</strong>
-                <small>{t(hintKey)}</small>
-              </span>
-            </label>
-          </li>
-        ))}
-      </ul>
-      {state === 'error' ? <p className="studio-error">{t('studioPanel.autonomy.failed')}</p> : null}
-      {state === 'saved' ? <p className="studio-autonomy-saved">{t('studioPanel.autonomy.saved')}</p> : null}
-    </div>
-  );
-}
-
-/** Ordered least to most permission, so the list reads as a scale rather than a menu. */
-const AUTONOMY_CHOICES: Array<[AutonomyMode, string, string]> = [
-  ['digest-only', 'studioPanel.autonomy.digestOnly', 'studioPanel.autonomy.digestOnlyHint'],
-  ['suggest', 'studioPanel.autonomy.suggest', 'studioPanel.autonomy.suggestHint'],
-  ['auto-fix-defects', 'studioPanel.autonomy.autoFixDefects', 'studioPanel.autonomy.autoFixDefectsHint'],
-  ['auto-tune', 'studioPanel.autonomy.autoTune', 'studioPanel.autonomy.autoTuneHint'],
-];
