+++ apps/web/src/surfaces/studio/studioHealth.ts
@@ -1,0 +1,14 @@
+import type { GameHealth } from '../../healthApi.js';
+import type { StudioGame } from '../../studioApi.js';
+
+export function formatSeconds(seconds: number): string {
+  if (seconds < 60) return `${Math.round(seconds)}s`;
+  const minutes = Math.floor(seconds / 60);
+  const rest = Math.round(seconds % 60);
+  return rest === 0 ? `${minutes}m` : `${minutes}m ${rest}s`;
+}
+
+export function healthFor(game: StudioGame, rows: GameHealth[]): GameHealth | null {
+  if (!game.slug) return null;
+  return rows.find((row) => row.slug === game.slug) ?? null;
+}
